CDHSH-94 Refactored the ProcessedJiraData into its own file and cleaned it up.
